# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "MJ Ma",
            "url": "https://www.linkedin.com/in/mj-profile/",
            "email": "m.mjw.ma@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/shellingford37/jump-jump/blob/master/LICENSE"
        }
    },
    "paths": {
        "/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取系统配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "获取系统配置信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetConfigAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/config/id-length": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新短链接 ID 设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "更新短链接 ID 设置",
                "parameters": [
                    {
                        "description": "更新短链接 ID 设置请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/IdConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetConfigAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/config/landing-hosts": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新落地页 Hosts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "更新落地页 Hosts",
                "parameters": [
                    {
                        "description": "更新落地页 Hosts 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateLandingHostsAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetConfigAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/config/short-link-404-handling": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新短链接 404 设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统配置"
                ],
                "summary": "更新短链接 404 设置",
                "parameters": [
                    {
                        "description": "更新短链接 404 设置请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ShortLinkNotFoundConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetConfigAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "GetConfigAPIResponseData": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/SystemConfig"
                }
            }
        },
...
        "IdConfig": {
            "type": "object",
            "properties": {
                "idLength": {
                    "description": "ID 长度",
                    "type": "integer",
                    "format": "int",
                    "example": 6
                },
                "idMaximumLength": {
                    "description": "最大 ID 长度",
                    "type": "integer",
                    "format": "int",
                    "example": 10
                },
                "idMinimumLength": {
                    "description": "最小 ID 长度",
                    "type": "integer",
                    "format": "int",
                    "example": 2
                }
            }
        },
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int",
                    "default": 0,
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string",
                    "default": "ok",
                    "example": "ok"
                }
            }
        },
...
        "ShortLinkNotFoundConfig": {
            "type": "object",
            "required": [
                "mode",
                "value"
            ],
            "properties": {
                "mode": {
                    "description": "模式",
                    "type": "string",
                    "enum": [
                        "content",
                        "redirect"
                    ],
                    "example": "content"
                },
                "value": {
                    "description": "值",
                    "type": "string",
                    "example": "page not found"
                }
            }
        },
        "SystemConfig": {
            "type": "object",
            "properties": {
                "idConfig": {
                    "description": "ID 配置",
                    "$ref": "#/definitions/IdConfig"
                },
                "landingHosts": {
                    "description": "落地页 Host 列表",
                    "type": "array",
                    "format": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://a.com/",
                        "https://b.com/"
                    ]
                },
                "shortLinkNotFoundConfig": {
                    "description": "短链接 404 配置",
                    "$ref": "#/definitions/ShortLinkNotFoundConfig"
                }
            }
        },
        "UpdateLandingHostsAPIRequest": {
            "type": "object",
            "properties": {
                "hosts": {
                    "type": "array",
                    "format": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://a.com/",
                        "https://b.com/"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}